feat: add suggestion search on homescreen, separate component for suggestion search
