+++ src/sass/main.scss
@@ -15,6 +15,7 @@
 // Components
 @import 'components/buttons';
 @import 'components/card';
+@import 'components/suggestions';
 
 // Pages
 @import 'pages/home';
